+++ CHANGELOG.md
@@ -7,6 +7,17 @@ Unreleased
 * Added ApplicationPageStatus
 * Added ApplicationPageStatusProvider to ApplicationBase, ApplicationNavigation, ModalManager and SlidePanelManager
 
+1.26.0 - (June 16, 2020)
+------------------
+### Added
+* Added a logger utility.
+
+1.25.0 - (June 9, 2020)
+------------------
+### Changed
+* Updated eslint-config-terra to @cerner/eslint-config-terra v^4.0.0
+* Removed unnecessary disabled rules.
+
 1.24.0 - (May 12, 2020)
 ------------------
 ### Added
@@ -20,6 +31,7 @@ Unreleased
 1.22.0 - (April 28, 2020)
 ------------------
 ### Changed
+* Updated errorBoundary to have a default error message when there is none provided.
 * Update terra-toolkit dev dependency to 6.0.0
 
 1.21.0 - (April 22, 2020)